Made great changes to the player manager and added a new file for the player manager
